startup scripts renamed and acceptance tests added
