--- src/bootstrap/download_key
@@ -1,51 +1,0 @@
------BEGIN RSA PRIVATE KEY-----
-MIIJKgIBAAKCAgEAwKE9fQU8dvRRdU+4NhANj8nOEyg6JhhogrU4js4d4rb7f2tk
-yQadEi3EAgX26apX9nxgXNhQfbNo6GoQTy042tHroMIYuAj+Bqg2AK+j9WP+KRxi
-9SbQSR48mUfSxS0GJxK4gPKRHD+GpvfGtsLdMYZJoNFWAJ8glxlzHT9hwNvoBbnZ
-x3m8tsQx/uvy0CBVS4pD/CnNmKH+Nbi6EAl5lxDTorq8zU4HHuG5Ufp+nbFG6f7q
-V1ZPk576u1OeCKPOUVCPU5BUR8CDl0mDe8HS1wtOm1h/FEqq+64/fn6Sk1fQLtQs
-efzFGvYxFHONb1iDtTreEhhQeiXmuTwxyMN/545F3huOQqk+pGqMhF5WLubkYcZD
-B1fY+zJTqJunfvtxmUFJaQHkaFNX3iACwT0EKDNulSLAbuF5p9FHKx3IY6BxAtlo
-uYelOY7ofCCwnUkqXtmMrz6qN9gXy8kkndeCRstm6RSFC0yih6YNSIaM8dKaEF4j
-rRJ3ifW+G6FeIJ39KKr8wvSpQ1BATLR0Z0FzAnDqlNXt9ZDRgEAlaEz5Em/ZR7nn
-JVpB6qTNZ4UBhhh5vO7RElPtBglngupzcAoRm7Nphy5s4hk78aFJwGAUUSCuY4vy
-4qiRfG9Sc+UJu2msiQ5v+eLsvHx58tNo8Xy6nUA3V5ZXMp1JAGfbDNBw+s8CAwEA
-AQKCAgBO6n/HVTrVPPiWR+ee9mqn6rQrMwLHkp5WVinqj1S9Cdv08qlilfwJc4hm
-j8z9I5VP3NqZjuFC7hpfhPxTg6JUO0X+ISaBVuXDEmy1UT+pi6KXcxKzfog8PE/1
-N/CIaQ6pOUkvQrDiUsA2bTISs7hszqghY0GJlaeDX6DCbbH9kdn3ZpFq4FvIwI3+
-8ukSDP0Eqot5piKdklgT6C8J3OfcVL9vXnN05K+UHkOO60GuFu+Hc7DLwQdu4eW0
-kugegAVJWOivAmUU1DDU7QAZ1mkrQm2cxN+jD4gG/fbzOqWQCZy3YqMqAr05c3zd
-9Q0PfcLPn2NVJHLr59UjNB7tMfKW0mmD4OZsgYxQ9emrbil3Q3L4nv8bZ/nz/JQL
-RcVDWcrx5aYtXcxQhOmvltyI0cmj/UxfqCF8uGsudWm5Crw9/Mqij5O/5VeL2rm7
-V1HeH8jVuJTg+HGLo977feVIjpk/xDakXaBNOVTUGeWk1RWQs+XVBjgw79Q9NG/2
-O/cecu05phBmROeslElcpvFdQ+8uL8bfuzvJLz1E4kZf9Plf84qJN5I9lUMW3mjU
-S0DoUaJ942Xjm9l/rnFazAtqgdbSu26sEZemnWji1ObjaK2bvx5cmRFSwLIDugIa
-N71WD6iRWpIahwOH/obCeUElS07KkbuGzSaquwiKEv+zYvjBuQKCAQEA+uwdQXSn
-ccKhZW57XFMYUIvNJFMYmaSabM+XTeCPOvecTzA1cbNACGiJGvxyQAj0DAd0tXpS
-lPrZ0YvzhE/vsIwpiWc25DxNPLbnCEkOZuyPl3/G/zOwAxnwKPXnRvyv9mTeMNk0
-mncAPBdrGbBJqbJNdtI0qCObRwieQkQ0ury3p5MHQS26w4xrDWIr3H8YgmLDScWA
-ReDh5p0Qx/4NIuKeGZ4OnvUU/DsCGsSiqoAWUvNzZHmf+43yxQFRbAom2PH/u5hV
-n+zVQohGQpWzomEG2ZH3GuWALFx0jDf6SzVnr4enDPj7xbPp5kk0eNPp6AFHR/Gb
-yl/T5xSRP7YJ7QKCAQEAxIclUPVBebSRQg7LmIUcPgNqQCuAkUhfc1+M+0bflv9O
-gqLHa3KrA0KCS5JNNi+K2r/DPQbA0DDlntsgNz2Adyr26BLUniIFO6D1NKd3OMBJ
-4ZTA/XLZb7nfDMRZGwHX4jW0xLcDzwMni0cHHS9wqa1j11EY1pgKBZaoRO8zmToC
-b3hzpzipQ4/NZdV72McA8v4ERYnQaDeL+wL3gCmNH+9+ptdRaeBCsZqs2VXiGV15
-qa8LEnO3v7Y3G0BEMQiazxtrCEWeuRBBTOs3qWh9i3pvklU2jdObp5o5CLZQqmH0
-FD7hLgPbZEzC/QM5WrSKBAcIbv3rGGhPVow9i2iQKwKCAQEA8zr+AP07cyLfoJtA
-KXbDExDxgkyqhATzTJbObEdJkDeF5Z0tsVG4uG3cubQMa9ZB+4cesuwoMRIZIu8u
-1wtFpTW47eCHbSanGDuUKIvo0gxuBMVv+OH5tIfoLUxxLHa1e3q7h06Fl1XTfGyW
-/QfN4BlezddmA7bARlMsVcAX/QwCWwZC+gOCDxyGiSqoqczpzTo/Y7t3ZEVhtVJ7
-eeHiifZymxKDycJ43Se2GoyRwgT3nuEu8i6Tzh0DP8TPUJNmpNWwuAAqwCOrdhrz
-QIgs4TJ8mVrM1d2yKU5WC7YXgd/66kPswr54EGUeToSXwKhrDOJwXmTuR3Ysyss9
-IS0s4QKCAQEAlAwjBlvbmS13h8XRhfIwll7IOdMLpWZZ0XYXpGt9oChwImUxyTW2
-slv1/6wU0k76UVPIZj6iHbPYJtV37WU9U4hgO++DkCScwJC+Z44IeA4TGgaPY96T
-AmGJL/ryDTBUSyqeJNqh2d7MeVVmfIvyWUAvSFLlXTNnyVH3LQbtXe6fSw34Qe6j
-RCweVSp/g4xWoLbOzPvGrdu6bbuMa1Xh5D7ssA7gW9v/IgAdWhQioW0Ta8pbkujt
-cVR5KuQ8+ps60+4ZsT+RFGS0bJd9dPcsLFhtmFTtJeZVCetaL7/4AIKN9npv89Qf
-/+z7rtIw8K20ktgheEt8QQTDT7WR32VDxQKCAQEAx0QCvXXFwxs+unz9H+x7Pequ
-NgkfWElYvuRM0VPBgg2379RW8E/0UYpD6BtDgg1LjVBQe16lOQvUMdekwL6pEgfl
-+GyPiWC8plrEFiW/eJl2M1H64nDC7s04GBhldfTY4EBTgiApvvpap0W1TISH8FyN
-WTQT0KqxLoPDIRvSXFKE7s5VmXLnWezrJ2QWlnHGNdpaFV+WTLiCqyLb1IXVsdMA
-iMDgKngWk7lFds8Ll1No5qUj+YzdH9dqfupwhsSOjSJNFw0xhHmzsrVloAvVJTRz
-AXs6qcn+7Vd3oiT/IaNRl+G2oMO81ZY0gGnMLS4/jV4BVnHwPfYsSAv2EnPjJQ==
------END RSA PRIVATE KEY-----
