--- src/bootstrap/download_key.pub
@@ -1,1 +1,0 @@
-ssh-rsa AAAAB3NzaC1yc2EAAAADAQABAAACAQDAoT19BTx29FF1T7g2EA2Pyc4TKDomGGiCtTiOzh3itvt/a2TJBp0SLcQCBfbpqlf2fGBc2FB9s2joahBPLTja0eugwhi4CP4GqDYAr6P1Y/4pHGL1JtBJHjyZR9LFLQYnEriA8pEcP4am98a2wt0xhkmg0VYAnyCXGXMdP2HA2+gFudnHeby2xDH+6/LQIFVLikP8Kc2Yof41uLoQCXmXENOiurzNTgce4blR+n6dsUbp/upXVk+Tnvq7U54Io85RUI9TkFRHwIOXSYN7wdLXC06bWH8USqr7rj9+fpKTV9Au1Cx5/MUa9jEUc41vWIO1Ot4SGFB6Jea5PDHIw3/njkXeG45CqT6kaoyEXlYu5uRhxkMHV9j7MlOom6d++3GZQUlpAeRoU1feIALBPQQoM26VIsBu4Xmn0UcrHchjoHEC2Wi5h6U5juh8ILCdSSpe2YyvPqo32BfLySSd14JGy2bpFIULTKKHpg1Ihozx0poQXiOtEneJ9b4boV4gnf0oqvzC9KlDUEBMtHRnQXMCcOqU1e31kNGAQCVoTPkSb9lHueclWkHqpM1nhQGGGHm87tESU+0GCWeC6nNwChGbs2mHLmziGTvxoUnAYBRRIK5ji/LiqJF8b1Jz5Qm7aayJDm/54uy8fHny02jxfLqdQDdXllcynUkAZ9sM0HD6zw== install_key@faf
